Renamed TextDoc => Document, TextCorpus => Corpus
also split the classes into separate modules
