+++ tests/test_readme.py
@@ -12,7 +12,7 @@ class ReadmeTestCase(unittest.TestCase):
 
     def setUp(self):
         self.text = "Mr. President, I ask to have printed in the Record copies of some of the finalist essays written by Vermont High School students as part of the sixth annual ``What is the State of the Union'' essay contest conducted by my office. These finalists were selected from nearly 800 entries. The material follows: The United States of America is an amazing nation that continues to lead the world through the complex geopolitical problems that we are faced with today. As a strong economic and political world leader, we have become the role model for developing nations attempting to give their people the same freedoms and opportunities that Americans have become so accustomed to. This is why it is so important to work harder than we ever have before to better ourselves as a nation, because what we change will set a precedent of improvement around the world and inspire change. The biggest problem in the U.S. is the incarceration system. It has been broken for decades, and there has been no legitimate attempt to fix it. Over the past thirty years, there has been a 500% increase in incarceration rates, resulting in the U.S. leading the world in number of prisoners with 2.2 million people currently incarcerated. Especially in this example, it is important to humanize these statistics. These are 2.2 million people, who now because of their conviction will find it much harder to be truly integrated back in their communities, due to the struggles of finding a job with a record, and the fact that they often do not qualify for social welfare. The incarceration system is also bankrupting both the state and federal government. It currently is the third highest state expenditure, behind health care and education. Fortunately, we as a nation have the opportunity to fix the incarceration system. First, we need to get rid of mandatory minimum sentences. Judges from across the nation have said for decades that they do not like mandatory minimums, that they do not work, and that they are unconstitutional. Mandatory minimum sentences, coupled with racially biased laws concerning drug possession is the reason why we see the ratio of African American males to white males over 10:1. This leads to the second action we must take; we must end the war on drugs. It has proven to be a failed experiment that has reopened many racial wounds in our nation. The war on drugs also put addicts behind bars, rather than treating addiction like the problem it actually is; a mental health issue."
-        self.doc = textacy.TextDoc(self.text)
+        self.doc = textacy.Document(self.text)
         cols = [attrs.TAG, attrs.HEAD, attrs.DEP]
         values = np.array(
             [[441, 1, 9480], [441, 3, 392], [416, 2, 407], [445, 1, 393],
