+++ tests/test_representations_vsm.py
@@ -6,7 +6,7 @@ import unittest
 import numpy as np
 from scipy.sparse import coo_matrix
 
-from textacy.texts import TextCorpus
+from textacy import Corpus
 from textacy.representations import vsm
 
 
@@ -16,7 +16,7 @@ class RepresentationsVSMTestCase(unittest.TestCase):
         texts = ["Burton loves to work with data — especially text data.",
                  "Extracting information from unstructured text is an interesting challenge.",
                  "Sometimes the hardest part is acquiring the right text; as with much analysis, it's garbage in, garbage out."]
-        textcorpus = TextCorpus.from_texts('en', texts)
+        textcorpus = Corpus.from_texts('en', texts)
         term_lists = [doc.as_terms_list(words=True, ngrams=False, named_entities=False)
                       for doc in textcorpus]
         self.doc_term_matrix, self.id_to_word = vsm.build_doc_term_matrix(
