+++ textacy/__init__.py
@@ -18,11 +18,11 @@ from textacy import data, preprocess, text_utils
 from textacy import lexicon_methods, spacy_utils, text_stats
 from textacy import extract
 from textacy import distance, export, keyterms
-from textacy import texts
 
 from textacy.data import load_spacy
 from textacy.preprocess import preprocess_text
-from textacy.texts import TextDoc, TextCorpus
+from textacy.document import Document
+from textacy.corpus import Corpus
 
 logger = logging.getLogger('textacy')
 if len(logger.handlers) == 0:  # To ensure reload() doesn't add another one
