+++ textacy/corpora/capitolwords.py
@@ -72,11 +72,11 @@ class CapitolWords(object):
         >>> for text in cw.texts(min_len=50000):
         ...     print(len(text))
 
-    Stream speeches into a `TextCorpus`::
+    Stream speeches into a `Corpus`::
 
         >>> text_stream, metadata_stream = textacy.fileio.split_content_and_metadata(
         ...     cw.docs(limit=100), 'text')
-        >>> tc = textacy.TextCorpus.from_texts('en', text_stream, metadata_stream)
+        >>> tc = textacy.Corpus.from_texts('en', text_stream, metadata_stream)
         >>> print(tc)
 
     Args:
