+++ textacy/document.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
 """
-Object-oriented interface for processing individual text documents as well as
-collections (corpora).
+Load, process, iterate, transform, and save text content paired with metadata —
+a document.
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from collections import Counter
-import copy
 import os
 import re
 import warnings
@@ -17,26 +16,26 @@ from spacy.tokens.doc import Doc as sdoc
 from spacy.tokens.token import Token as stoken
 from spacy.tokens.span import Span as sspan
 
-from textacy.compat import PY2, string_types, unicode_type, zip
+from textacy.compat import string_types, unicode_type
 from textacy import (data, extract, fileio, keyterms, spacy_utils,
                      text_stats, text_utils)
-from textacy.representations import network, vsm
+from textacy.representations import network
 
 
-class TextDoc(object):
+class Document(object):
     """
     Pairing of the content of a document with its associated metadata, where the
     content has been tokenized, tagged, parsed, etc. by spaCy. Also keep references
     to the id-to-token mapping used by spaCy to efficiently represent the content.
     If initialized from plain text, this processing is performed automatically.
 
-    ``TextDoc`` also provides a convenient interface to information extraction,
+    ``Document`` also provides a convenient interface to information extraction,
     different doc representations, statistical measures of the content, and more.
 
     Args:
         text_or_sdoc (str or ``spacy.Doc``): text or spacy doc containing the
-            content of this ``TextDoc``; if str, it is automatically processed
-            by spacy before assignment to the ``TextDoc.spacy_doc`` attribute
+            content of this ``Document``; if str, it is automatically processed
+            by spacy before assignment to the ``Document.spacy_doc`` attribute
         spacy_pipeline (``spacy.<lang>.<Lang>()``, optional): if a spacy pipeline
             has already been loaded or used to make the input ``spacy.Doc``, pass
             it in here to speed things up a touch; in general, only one of these
@@ -61,7 +60,7 @@ class TextDoc(object):
             # check for match between text and passed spacy_pipeline language
             else:
                 if spacy_pipeline.lang != self.lang:
-                    msg = 'TextDoc.lang {} != spacy_pipeline.lang {}'.format(
+                    msg = 'Document.lang {} != spacy_pipeline.lang {}'.format(
                         self.lang, spacy_pipeline.lang)
                     raise ValueError(msg)
             self.spacy_vocab = spacy_pipeline.vocab
@@ -76,7 +75,7 @@ class TextDoc(object):
             self.spacy_doc = text_or_sdoc
 
         else:
-            msg = 'TextDoc must be initialized with {}, not {}'.format(
+            msg = 'Document must be initialized with {}, not {}'.format(
                 {str, sdoc}, type(text_or_sdoc))
             raise ValueError(msg)
 
@@ -84,7 +83,7 @@ class TextDoc(object):
         snippet = self.text[:50].replace('\n', ' ')
         if len(snippet) == 50:
             snippet = snippet[:47] + '...'
-        return 'TextDoc({} tokens; "{}")'.format(len(self.spacy_doc), snippet)
+        return 'Document({} tokens; "{}")'.format(len(self.spacy_doc), snippet)
 
     def __len__(self):
         return self.n_tokens
@@ -98,7 +97,7 @@ class TextDoc(object):
 
     def save(self, path, fname_prefix=None):
         """
-        Save serialized TextDoc content and metadata to disk.
+        Save serialized Document content and metadata to disk.
 
         Args:
             path (str): directory on disk where content + metadata will be saved
@@ -124,7 +123,7 @@ class TextDoc(object):
     @classmethod
     def load(cls, path, fname_prefix=None):
         """
-        Load serialized content and metadata from disk, and initialize a TextDoc.
+        Load serialized content and metadata from disk, and initialize a Document.
 
         Args:
             path (str): directory on disk where content + metadata are saved
@@ -133,7 +132,7 @@ class TextDoc(object):
                 when saving to disk
 
         Returns:
-            :class:`textacy.TextDoc`
+            :class:`textacy.Document`
 
         .. warn:: If the `spacy.Vocab` object used to save this document is not the
             same as the one used to load it, there will be problems! Consequently,
@@ -150,10 +149,10 @@ class TextDoc(object):
         spacy_version = metadata.pop('spacy_version')
         if spacy_version != spacy.about.__version__:
             msg = """
-                the spaCy version used to save this TextDoc to disk is not the
+                the spaCy version used to save this Document to disk is not the
                 same as the version currently installed ('{}' vs. '{}'); if the
                 data underlying the associated `spacy.Vocab` has changed, this
-                loaded TextDoc may not be valid!
+                loaded Document may not be valid!
                 """.format(spacy_version, spacy.about.__version__)
             warnings.warn(msg, UserWarning)
         spacy_vocab = data.load_spacy(lang).vocab
@@ -219,7 +218,7 @@ class TextDoc(object):
             weighting (str {'tf', 'tfidf'}, optional): weighting of term weights,
                 either term frequency ('tf') or tf * inverse doc frequency ('tfidf')
             idf (dict, optional): if `weighting` = 'tfidf', idf's must be supplied
-                externally, such as from a `TextCorpus` object
+                externally, such as from a `Corpus` object
             lemmatize (bool or 'auto', optional): if True, lemmatize all terms
                 when getting their frequencies
             ngram_range (tuple(int), optional): (min n, max n) values for n-grams
@@ -725,337 +724,3 @@ class TextDoc(object):
     @property
     def readability_stats(self):
         return text_stats.readability_stats(self)
-
-
-class TextCorpus(object):
-    """
-    An ordered collection of :class:`TextDoc <textacy.texts.TextDoc>` s, all of
-    the same language and sharing a single spaCy pipeline and vocabulary. Tracks
-    overall corpus statistics and provides a convenient interface to alternate
-    corpus representations.
-
-    Args:
-        lang_or_pipeline ({'en', 'de'} or :class:`spacy.<lang>.<Language>`):
-            either 'en' or 'de', used to initialize the corresponding spaCy pipeline
-            with all models loaded by default, or an already-initialized spaCy
-            pipeline which may or may not have all models loaded
-
-    Add texts to corpus one-by-one with :meth:`TextCorpus.add_text() <textacy.texts.TextCorpus.add_text>`,
-    or all at once with :meth:`TextCorpus.from_texts() <textacy.texts.TextCorpus.from_texts>`.
-    Can also add already-instantiated TextDocs via :meth:`TextCorpus.add_doc() <textacy.texts.TextCorpus.add_doc>`.
-
-    Iterate over corpus docs with ``for doc in TextCorpus``. Access individual docs
-    by index (e.g. ``TextCorpus[0]`` or ``TextCorpus[0:10]``) or by boolean condition
-    specified by a function (e.g. ``TextCorpus.get_docs(lambda x: len(x) > 100)``).
-    """
-    def __init__(self, lang_or_pipeline):
-        if isinstance(lang_or_pipeline, string_types):
-            self.lang = lang_or_pipeline
-            self.spacy_pipeline = data.load_spacy(self.lang)
-        else:
-            self.spacy_pipeline = lang_or_pipeline
-            self.lang = self.spacy_pipeline.lang
-        self.spacy_vocab = self.spacy_pipeline.vocab
-        self.spacy_stringstore = self.spacy_vocab.strings
-        self.docs = []
-        self.n_docs = 0
-        self.n_sents = 0
-        self.n_tokens = 0
-
-    def __repr__(self):
-        return 'TextCorpus({} docs; {} tokens)'.format(self.n_docs, self.n_tokens)
-
-    def __len__(self):
-        return self.n_docs
-
-    def __getitem__(self, index):
-        return self.docs[index]
-
-    def __delitem__(self, index):
-        del self.docs[index]
-
-    def __iter__(self):
-        for doc in self.docs:
-            yield doc
-
-    def save(self, path, fname_prefix=None, compression=None):
-        """
-        Save serialized TextCorpus content and metadata to disk.
-
-        Args:
-            path (str): directory on disk where content + metadata will be saved
-            fname_prefix (str, optional): prepend standard filenames 'spacy_docs.bin'
-                and 'metadatas.json' with additional identifying information
-            compression ({'gzip', 'bz2', 'lzma'} or None): type of compression
-                used to reduce size of metadatas json file
-
-        .. warn:: If the `spacy.Vocab` object used to save this corpus is not the
-            same as the one used to load it, there will be problems! Consequently,
-            this functionality is only useful as short-term but not long-term storage.
-        """
-        if fname_prefix:
-            info_fname = os.path.join(path, '_'.join([fname_prefix, 'info.json']))
-            meta_fname = os.path.join(path, '_'.join([fname_prefix, 'metadatas.json']))
-            docs_fname = os.path.join(path, '_'.join([fname_prefix, 'spacy_docs.bin']))
-        else:
-            info_fname = os.path.join(path, 'info.json')
-            meta_fname = os.path.join(path, 'metadatas.json')
-            docs_fname = os.path.join(path, 'spacy_docs.bin')
-        meta_fname = meta_fname + ('.gz' if compression == 'gzip'
-                                   else '.bz2' if compression == 'bz2'
-                                   else '.xz' if compression == 'lzma'
-                                   else '')
-        meta_mode = 'wt' if PY2 is False or compression is None else 'wb'
-        package_info = {'textacy_lang': self.lang, 'spacy_version': spacy.about.__version__}
-        fileio.write_json(package_info, info_fname)
-        fileio.write_json_lines(
-            (doc.metadata for doc in self), meta_fname, mode=meta_mode,
-            ensure_ascii=False, separators=(',', ':'))
-        fileio.write_spacy_docs((doc.spacy_doc for doc in self), docs_fname)
-
-    @classmethod
-    def load(cls, path, fname_prefix=None, compression=None):
-        """
-        Load serialized content and metadata from disk, and initialize a TextCorpus.
-
-        Args:
-            path (str): directory on disk where content + metadata are saved
-            fname_prefix (str, optional): additional identifying information
-                prepended to standard filenames 'spacy_docs.bin' and 'metadatas.json'
-                when saving to disk
-            compression ({'gzip', 'bz2', 'lzma'} or None): type of compression
-                used to reduce size of metadatas json file
-
-        Returns:
-            :class:`textacy.TextCorpus`
-
-        .. warn:: If the `spacy.Vocab` object used to save this corpus is not the
-            same as the one used to load it, there will be problems! Consequently,
-            this functionality is only useful as short-term but not long-term storage.
-        """
-        if fname_prefix:
-            info_fname = os.path.join(path, '_'.join([fname_prefix, 'info.json']))
-            meta_fname = os.path.join(path, '_'.join([fname_prefix, 'metadatas.json']))
-            docs_fname = os.path.join(path, '_'.join([fname_prefix, 'spacy_docs.bin']))
-        else:
-            info_fname = os.path.join(path, 'info.json')
-            meta_fname = os.path.join(path, 'metadatas.json')
-            docs_fname = os.path.join(path, 'spacy_docs.bin')
-        meta_fname = meta_fname + ('.gz' if compression == 'gzip'
-                                   else '.bz2' if compression == 'bz2'
-                                   else '.xz' if compression == 'lzma'
-                                   else '')
-        meta_mode = 'rt' if PY2 is False or compression is None else 'rb'
-        package_info = list(fileio.read_json(info_fname))[0]
-        lang = package_info['textacy_lang']
-        spacy_version = package_info['spacy_version']
-        if spacy_version != spacy.about.__version__:
-            msg = """
-                the spaCy version used to save this TextCorpus to disk is not the
-                same as the version currently installed ('{}' vs. '{}'); if the
-                data underlying the associated `spacy.Vocab` has changed, this
-                loaded TextCorpus may not be valid!
-                """.format(spacy_version, spacy.about.__version__)
-            warnings.warn(msg, UserWarning)
-        textcorpus = TextCorpus(lang)
-        metadata_stream = fileio.read_json_lines(meta_fname, mode=meta_mode,)
-        spacy_docs = fileio.read_spacy_docs(textcorpus.spacy_vocab, docs_fname)
-        for spacy_doc, metadata in zip(spacy_docs, metadata_stream):
-            textcorpus.add_doc(
-                TextDoc(spacy_doc, spacy_pipeline=textcorpus.spacy_pipeline,
-                        lang=lang, metadata=metadata))
-        return textcorpus
-
-    @classmethod
-    def from_texts(cls, lang_or_pipeline, texts, metadata=None, n_threads=2, batch_size=1000):
-        """
-        Convenience function for creating a :class:`TextCorpus <textacy.texts.TextCorpus>`
-        from an iterable of text strings.
-
-        Args:
-            lang_or_pipeline ({'en', 'de'} or :class:`spacy.<lang>.<Language>`)
-            texts (iterable(str))
-            metadata (iterable(dict), optional)
-            n_threads (int, optional)
-            batch_size (int, optional)
-
-        Returns:
-            :class:`TextCorpus <textacy.texts.TextCorpus>`
-        """
-        textcorpus = cls(lang_or_pipeline)
-        spacy_docs = textcorpus.spacy_pipeline.pipe(
-            texts, n_threads=n_threads, batch_size=batch_size)
-        if metadata is not None:
-            for spacy_doc, md in zip(spacy_docs, metadata):
-                textcorpus.add_doc(TextDoc(spacy_doc, lang=textcorpus.lang,
-                                           spacy_pipeline=textcorpus.spacy_pipeline,
-                                           metadata=md))
-        else:
-            for spacy_doc in spacy_docs:
-                textcorpus.add_doc(TextDoc(spacy_doc, lang=textcorpus.lang,
-                                           spacy_pipeline=textcorpus.spacy_pipeline,
-                                           metadata=None))
-        return textcorpus
-
-    def add_text(self, text, lang=None, metadata=None):
-        """
-        Create a :class:`TextDoc <textacy.texts.TextDoc>` from ``text`` and ``metadata``,
-        then add it to the corpus.
-
-        Args:
-            text (str): raw text document to add to corpus as newly instantiated
-                :class:`TextDoc <textacy.texts.TextDoc>`
-            lang (str, optional):
-            metadata (dict, optional): dictionary of document metadata, such as::
-
-                {"title": "My Great Doc", "author": "Burton DeWilde"}
-
-                NOTE: may be useful for retrieval via :func:`get_docs() <textacy.texts.TextCorpus.get_docs>`,
-                e.g. ``TextCorpus.get_docs(lambda x: x.metadata["title"] == "My Great Doc")``
-        """
-        doc = TextDoc(text, spacy_pipeline=self.spacy_pipeline,
-                      lang=lang, metadata=metadata)
-        doc.corpus_index = self.n_docs
-        doc.corpus = self
-        self.docs.append(doc)
-        self.n_docs += 1
-        self.n_tokens += doc.n_tokens
-        # sentence segmentation requires parse; if not available, just skip this
-        try:
-            self.n_sents += doc.n_sents
-        except ValueError:
-            pass
-
-    def add_doc(self, doc, print_warning=True):
-        """
-        Add an existing :class:`TextDoc <textacy.texts.TextDoc>` to the corpus as-is.
-        NB: If ``textdoc`` is already added to this or another :class:`TextCorpus <textacy.texts.TextCorpus>`,
-        a warning message will be printed and the ``corpus_index`` attribute will be
-        overwritten, but you won't be prevented from adding the doc.
-
-        Args:
-            doc (:class:`TextDoc <textacy.texts.TextDoc>`)
-            print_warning (bool, optional): if True, print a warning message if
-                ``doc`` already added to a corpus; otherwise, don't ever print
-                the warning and live dangerously
-        """
-        if doc.lang != self.lang:
-            msg = 'TextDoc.lang {} != TextCorpus.lang {}'.format(doc.lang, self.lang)
-            raise ValueError(msg)
-        if hasattr(doc, 'corpus_index'):
-            doc = copy.deepcopy(doc)
-            if print_warning is True:
-                print('**WARNING: TextDoc already associated with a TextCorpus; adding anyway...')
-        doc.corpus_index = self.n_docs
-        doc.corpus = self
-        self.docs.append(doc)
-        self.n_docs += 1
-        self.n_tokens += doc.n_tokens
-        # sentence segmentation requires parse; if not available, just skip this
-        try:
-            self.n_sents += doc.n_sents
-        except ValueError:
-            pass
-
-    def get_doc(self, index):
-        """
-        Get a single doc by its position ``index`` in the corpus.
-        """
-        return self.docs[index]
-
-    def get_docs(self, match_condition, limit=None):
-        """
-        Iterate over all docs in corpus and return all (or N = ``limit``) for which
-        ``match_condition(doc) is True``.
-
-        Args:
-            match_condition (func): function that operates on a :class:`TextDoc`
-                and returns a boolean value; e.g. `lambda x: len(x) > 100` matches
-                all docs with more than 100 tokens
-            limit (int, optional): if not `None`, maximum number of matched docs
-                to return
-
-        Yields:
-            :class:`TextDoc <textacy.texts.TextDoc>`: one per doc passing ``match_condition``
-                up to ``limit`` docs
-        """
-        if limit is None:
-            for doc in self:
-                if match_condition(doc) is True:
-                    yield doc
-        else:
-            n_matched_docs = 0
-            for doc in self:
-                if match_condition(doc) is True:
-                    n_matched_docs += 1
-                    if n_matched_docs > limit:
-                        break
-                    yield doc
-
-    def remove_doc(self, index):
-        """Remove the document at ``index`` from the corpus, and decrement the
-        ``corpus_index`` attribute on all docs that come after it in the corpus."""
-        n_tokens_removed = self[index].n_tokens
-        try:
-            n_sents_removed = self[index].n_sents
-        except ValueError:
-            n_sents_removed = 0
-        del self[index]
-        # reset `corpus_index` attribute on docs higher up in the list
-        for doc in self[index:]:
-            doc.corpus_index -= 1
-        # also decrement the corpus doc/sent/token counts
-        self.n_docs -= 1
-        self.n_sents -= n_sents_removed
-        self.n_tokens -= n_tokens_removed
-
-    def remove_docs(self, match_condition, limit=None):
-        """
-        Remove all (or N = ``limit``) docs in corpus for which ``match_condition(doc) is True``.
-        Re-set all remaining docs' ``corpus_index`` attributes at the end.
-
-        Args:
-            match_condition (func): function that operates on a :class:`TextDoc <textacy.texts.TextDoc>`
-                and returns a boolean value; e.g. ``lambda x: len(x) > 100`` matches
-                all docs with more than 100 tokens
-            limit (int, optional): if not None, maximum number of matched docs
-                to remove
-        """
-        remove_indexes = sorted(
-            (doc.corpus_index
-             for doc in self.get_docs(match_condition, limit=limit)),
-            reverse=True)
-        n_docs_removed = len(remove_indexes)
-        n_sents_removed = 0
-        n_tokens_removed = 0
-        for index in remove_indexes:
-            n_tokens_removed += self[index].n_tokens
-            try:
-                n_sents_removed += self[index].n_sents
-            except ValueError:
-                pass
-            del self[index]
-        # now let's re-set the `corpus_index` attribute for all docs at once
-        for i, doc in enumerate(self):
-            doc.corpus_index = i
-        # also decrement the corpus doc/sent/token counts
-        self.n_docs -= n_docs_removed
-        self.n_sents -= n_sents_removed
-        self.n_tokens -= n_tokens_removed
-
-    def as_doc_term_matrix(self, terms_lists, weighting='tf',
-                           normalize=True, smooth_idf=True, sublinear_tf=False,
-                           min_df=1, max_df=1.0, min_ic=0.0, max_n_terms=None):
-        """
-        Transform corpus into a sparse CSR matrix, where each row i corresponds
-        to a doc, each column j corresponds to a unique term, and matrix values
-        (i, j) correspond to the tf or tf-idf weighting of term j in doc i.
-
-        .. seealso:: :func:`build_doc_term_matrix <textacy.representations.vsm.build_doc_term_matrix>`
-        """
-        self.doc_term_matrix, self.id_to_term = vsm.build_doc_term_matrix(
-            terms_lists, weighting=weighting,
-            normalize=normalize, smooth_idf=smooth_idf, sublinear_tf=sublinear_tf,
-            min_df=min_df, max_df=max_df, min_ic=min_ic,
-            max_n_terms=max_n_terms)
-        return (self.doc_term_matrix, self.id_to_term)
